+++ Presentation/PresentationOptions.swift
@@ -44,6 +44,9 @@ public extension PresentationOptions {
     /// Disable any presentation animations.
     static let unanimated = PresentationOptions()
 
+    /// Allow modal dismissal with swipe down gesture when the presented view controller is inside a navigation stack
+    static let allowSwipeDismissAlways = PresentationOptions()
+
     /// Default options used unless any options are explicity passed when presented: `[embedInNavigationController]`
     static let defaults: PresentationOptions = [embedInNavigationController]